% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{accuracy}
\alias{accuracy}
\title{accuracy}
\usage{
accuracy(actual, pred, best = FALSE, threshold = 0.5)
}
\value{
a list of a RF Object, AUC, predictions
}
\description{
accuracy
}
\examples{
\dontrun{
TODO
}

}
