% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{add_selected_but_no_bought}
\alias{add_selected_but_no_bought}
\title{add_selected_but_no_bought}
\usage{
add_selected_but_no_bought(
  data,
  from,
  to,
  by,
  min_customer = 500,
  max_customer = 1000,
  odd_new_customer = 5,
  mu_sku = 30,
  sigma_sku = 20,
  mu_qtty = 30,
  sigma_qtty = 20
)
}
\value{
a new dataset
}
\description{
add_selected_but_no_bought
}
\examples{
\dontrun{
TODO
}

}
