% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_score.R
\name{auc}
\alias{auc}
\title{Area Under the Curve : compute AUC between actual values and predictions}
\usage{
auc(actual, pred)
}
\arguments{
\item{actual}{: vector, actual values (target)}

\item{pred}{: vector, predicted probabilities}
}
\value{
a value
}
\description{
Area Under the Curve : compute AUC between actual values and predictions
}
