% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{count_keywords}
\alias{count_keywords}
\title{Find keywords onto the SKU labels.}
\usage{
count_keywords(labels, patterns = keywords)
}
\arguments{
\item{labels}{: \code{characters}. Characters to look patterns into.}

\item{patterns}{: \code{characters} pretty obvious}
}
\value{
a list of (data.table, a coverage indicator, summary of keywords)
}
\description{
Find keywords onto the SKU labels.
}
\examples{
\dontrun{
sample(unique(data$Description), 5000)
l = count_keywords(unique(data$Description))
}

}
