% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_agg_freq_cncl}
\alias{create_agg_freq_cncl}
\title{Second part of features computing.}
\usage{
create_agg_freq_cncl(dt)
}
\arguments{
\item{dt}{: \code{data.table}. A data.table obtained from a groupby on customers}
}
\value{
a data.table object
}
\description{
Second part of features computing.
}
\examples{
\dontrun{
TODO
}

}
