% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_features}
\alias{create_features}
\title{Compute features on a rolling windows}
\usage{
create_features(
  from = as.Date("2010/03/01"),
  to = as.Date("2011/12/01"),
  by = "month",
  windows = c("M-1", "M-2", "M-3"),
  kind = "lag"
)
}
\arguments{
\item{from}{: \code{Date}. starting date. Required}

\item{to}{: \code{Date}. end date. Required}

\item{by}{: \code{character}. increment of the sequence of Dates}

\item{data}{: \code{data.table}. Complete dataset.}
}
\value{
a data.table object
}
\description{
Compute features on a rolling windows
}
\examples{
\dontrun{
TODO
}

}
