% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_features_on_period}
\alias{create_features_on_period}
\title{Compute features on a specific targeted period}
\usage{
create_features_on_period(
  data,
  start_rep,
  end_rep,
  windows,
  kind = "lag",
  target = "factor"
)
}
\arguments{
\item{data}{: \code{data.table}. Complete dataset.}

\item{start_rep}{: \code{character}. Start of the targeted period. Character to be parsed as date (format YYYY-MM-DD)}

\item{end_rep}{: \code{character}. End of the targeted period. Character to be parsed as date (format YYYY-MM-DD)}

\item{windows}{: \code{integers}. windows=c("M-1", "M-2", "M-3") or windows=c("T-1", "T-2", "T-3")}

\item{kind}{: \code{character}. One of c("lapse", "cumulative").}

\item{target}{: \code{character}. One of c("factor", "amount").}
}
\value{
a data.table object
}
\description{
Compute features on a specific targeted period
}
\examples{
\dontrun{
start_rep = "2011-05-01"
TODO
}

}
