% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_subset_data}
\alias{create_subset_data}
\title{Subset dataset from a specified date of end for as long as specified (in months).}
\usage{
create_subset_data(data, to = "2011-10-01", window_months = 3)
}
\arguments{
\item{data}{: \code{data.talbe}. Complete dataset to extract period from.}

\item{to}{: \code{character}. Start of the targeted period. Character to be parsed as date (format YYYY-MM-DD)}

\item{window_months}{: \code{integer}. Size of period window.}
}
\value{
a data.table object
}
\description{
Subset dataset from a specified date of end for as long as specified (in months).
}
\examples{
\dontrun{
TODO
}

}
