% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_var_reponse_montant}
\alias{create_var_reponse_montant}
\title{Create response variable : amount of products purchased in the targeted month.}
\usage{
create_var_reponse_montant(
  data,
  start_rep = "2011-10-01",
  end_rep = "2011-10-31"
)
}
\arguments{
\item{data}{: \code{data.table}. A subset of retailer II dataset.}

\item{start_rep}{: \code{character}. Start of the targeted period. Character to be parsed as date (format YYYY-MM-DD)}

\item{end_rep}{: \code{character}. End of the targeted period. Character to be parsed as date (format YYYY-MM-DD)}
}
\value{
a data.table object
}
\description{
Create response variable : amount of products purchased in the targeted month.
}
\examples{
\dontrun{
TODO
}

}
