% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_score.R
\name{drift_score}
\alias{drift_score}
\title{Drift Score : model detecting drift between 
train and test data sets}
\usage{
drift_score(X_train, X_test)
}
\arguments{
\item{X_train}{: data.table, train data set}

\item{X_test}{: data.table, test data set}
}
\value{
list of performance indicators (of drift model)
}
\description{
Drift Score : model detecting drift between 
train and test data sets
}
