% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{import_dataset}
\alias{import_dataset}
\title{Import dataset}
\usage{
import_dataset(path = "/home/mmasson/data/mlops-wbr/uk-retailer-ii.xlsx")
}
\arguments{
\item{path}{: \code{character}. Path to xlsx file}
}
\value{
a data.table object
}
\description{
Import dataset
}
\examples{
\dontrun{
dt = import_dataset()
}


}
