% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{train_rf}
\alias{train_rf}
\title{Train a RandomForest model. Compute AUC.}
\usage{
train_rf(agg)
}
\arguments{
\item{agg}{: \code{data.table}. Features and target.}
}
\value{
a list of a RF Object, AUC, predictions
}
\description{
Train a RandomForest model. Compute AUC.
}
\examples{
\dontrun{
agg_4trim <- create_features_on_period(data, "2011-05-01", "2011-05-30", c("T-1", "T-2", "T-3", "T-4"), kind="lag")
# Distribution of VAR_REP 
agg_4trim[, .N/nrow(agg_4trim), VAR_REP]
# Train RF
res_4trim <- train_rf(agg_4trim)
# Importance
varImp(res_4trim$model)
plot(varImp(res_4trim$model), top=30)
res_4trim$auc
}

}
