% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{add_selected_but_no_bought}
\alias{add_selected_but_no_bought}
\title{add_selected_but_no_bought}
\usage{
add_selected_but_no_bought(
  data,
  from,
  to,
  by,
  min_customer = 500,
  max_customer = 1000,
  odd_new_customer = 5,
  mu_sku = 30,
  sigma_sku = 20,
  mu_qtty = 30,
  sigma_qtty = 20
)
}
\arguments{
\item{data}{: a dataset in which we will add fake lines}

\item{from}{: same as from in ?seq.Date}

\item{to}{: same as to in ?seq.Date}

\item{by}{: same as by in ?seq.Date}

\item{min_customer}{: min number of customers the can be altered each month}

\item{max_customer}{: max number of customers the can be altered each month}

\item{odd_new_customer}{: inverse of probability to add a new customer rather than altering an existing one}

\item{mu_sku}{: number of SKU by customer to be added follows rnorm(mu_sku, sigma_sku)}

\item{sigma_sku}{: number of SKU by customer to be added follows a rnom(mu_sku, sigma_sku)}

\item{mu_qtty}{: quantity added for each of SKU added follows rnorm(mu_qtty, sigma_qtty)}

\item{sigma_qtty}{: quantity added for each of SKU added follows rnorm(mu_qtty, sigma_qtty)}
}
\value{
a new dataset
}
\description{
add_selected_but_no_bought
}
\examples{
\dontrun{
#TODO
}

}
