% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_agg_prix_qty}
\alias{create_agg_prix_qty}
\title{First part of features computing. Compute the average basket and other features.}
\usage{
create_agg_prix_qty(sub_data_agg, all_customers)
}
\arguments{
\item{sub_data_agg}{: \code{data.table}. A subset of the complete dataset(Use create_subset_data)}

\item{all_customers}{: \code{data.table}. List of all the considered customers.}
}
\value{
a data.table object
}
\description{
First part of features computing. Compute the average basket and other features.
}
\examples{
\dontrun{
TODO
}

}
