% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{monitoring_main}
\alias{monitoring_main}
\title{monitoring_main}
\usage{
monitoring_main(
  data,
  X_train,
  start,
  end,
  model,
  kind_target = "factor",
  kind_agregates = "cumulative",
  depth_agregats = c("M-6"),
  compute_drift = TRUE,
  compute_perf_globale = TRUE,
  delay_update = "month",
  date_format = "\%Y-\%m-\%d"
)
}
\value{
a list of a data.table with scores, and a list of data.table corresponding to each features for each months
}
\description{
monitoring_main
}
\examples{
\dontrun{
TODO
}

}
