% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{train_xgboost}
\alias{train_xgboost}
\title{Train a RandomForest model. Compute AUC.}
\usage{
train_xgboost(agg)
}
\arguments{
\item{agg}{: \code{data.table}. Features and target.}
}
\value{
a list of a RF Object, AUC, predictions
}
\description{
Train a RandomForest model. Compute AUC.
}
\examples{
\dontrun{
TODO
}

}
